+++ sc_rest_scraper/__init__.py
@@ -9,5 +9,5 @@ municipal and county engineering consulting.
 License: MIT
 """
 
-__version__ = '1.4.0'
+__version__ = '1.5.0'
 __author__ = 'GeoGrab Contributors'